Revert "39 implement drag and drop for board"
