+++ package.json
@@ -16,7 +16,6 @@
     "prop-types": "^15.7.2",
     "query-string": "^7.0.0",
     "react": "^17.0.1",
-    "react-beautiful-dnd": "^13.1.0",
     "react-dom": "^17.0.1",
     "react-redux": "^7.2.3",
     "react-router": "^5.2.0",
