+++ public/manifest.json
@@ -1,6 +1,23 @@
 {
-  "short_name": "Snailboard",
-  "name": "Snailboard | Make life easier",
+  "short_name": "React Redux Saga App",
+  "name": "Create React Redux Saga App Sample",
+  "icons": [
+    {
+      "src": "favicon.ico",
+      "sizes": "64x64 32x32 24x24 16x16",
+      "type": "image/x-icon"
+    },
+    {
+      "src": "logo192.png",
+      "type": "image/png",
+      "sizes": "192x192"
+    },
+    {
+      "src": "logo512.png",
+      "type": "image/png",
+      "sizes": "512x512"
+    }
+  ],
   "start_url": ".",
   "display": "standalone",
   "theme_color": "#000000",
