+++ src/components/Header.js
@@ -1,11 +1,10 @@
 import React from 'react'
-import PropTypes from 'prop-types'
 import {
   AppBar,
   Button,
   Grid,
   IconButton,
-  Paper,
+  makeStyles,
   Toolbar,
   Typography,
 } from '@material-ui/core'
@@ -13,13 +12,38 @@ import HomeIcon from '@material-ui/icons/Home'
 import { NavLink } from 'react-router-dom'
 import Icon from '@material-ui/core/Icon'
 import { useSelector } from 'react-redux'
-import snailLogo from '../../static/img/snailboard.png'
-import { userSelector } from '../../containers/Auth/authSlice'
-import { useStyles } from './styles'
-import { PRIMARY_COLOR } from '../../core/values/colors'
+import { PRIMARY_COLOR } from '../core/values/colors'
+import snailLogo from '../static/img/snailboard.png'
+import { userSelector } from '../containers/Auth/authSlice'
 
-const Header = (props) => {
-  const { boardName } = props
+export const useStyles = makeStyles((theme) => ({
+  menuButton: {
+    marginRight: theme.spacing(2),
+    justifyContent: 'center',
+    justifyItems: 'center',
+  },
+  title: {
+    flexGrow: 1,
+    fontSize: 'medium',
+  },
+  appBar: {
+    background: PRIMARY_COLOR,
+  },
+  btnLink: {
+    color: '#fff',
+    fontSize: 'medium',
+    textTransform: 'none',
+  },
+  snailBtn: {
+    justifyContent: 'center',
+    justifyItems: 'center',
+    margin: '10px',
+    width: '95px',
+    height: '60px',
+  },
+}))
+
+const Header = () => {
   const user = useSelector(userSelector)
 
   const classes = useStyles()
@@ -44,26 +68,13 @@ const Header = (props) => {
                     <HomeIcon fontSize="large" />
                   </IconButton>
                 </Grid>
-                {boardName && (
-                  <>
-                    <Grid item xs={3}>
-                      <Typography>
-                        <Button className={classes.btnLink}>Boards</Button>
-                      </Typography>
-                    </Grid>
-                    <Grid item xs={4}>
-                      <Paper
-                        style={{
-                          textAlign: 'center',
-                          color: PRIMARY_COLOR,
-                        }}
-                        variant="outlined"
-                      >
-                        {boardName}
-                      </Paper>
-                    </Grid>
-                  </>
-                )}
+                <Grid item xs={3}>
+                  <NavLink exact to="/boards">
+                    <Typography>
+                      <Button className={classes.btnLink}>Boards</Button>
+                    </Typography>
+                  </NavLink>
+                </Grid>
               </Grid>
             </Grid>
             <Grid item xs={4} align="center">
@@ -112,10 +123,6 @@ const Header = (props) => {
       </AppBar>
     </>
   )
-}
-
-Header.propTypes = {
-  boardName: PropTypes.string,
 }
 
 export default Header
