--- src/containers/HomePage/SBCard/cardTypes.js
@@ -1,1 +1,0 @@
-export const CARD_TYPES = { BOARD: 'BOARD', TEAM: 'TEAM' }
