+++ src/containers/Modals/AddTeamModal.js
@@ -1,29 +1,30 @@
 import React, { useState } from 'react'
 
 import {
-  Backdrop,
-  Button,
-  CircularProgress,
   Grid,
-  makeStyles,
   Modal,
   Paper,
-  Snackbar,
-  TextareaAutosize,
-  TextField,
+  makeStyles,
   Typography,
+  TextField,
+  TextareaAutosize,
+  Button,
+  CircularProgress,
+  Backdrop,
+  Snackbar,
 } from '@material-ui/core'
+import PropTypes from 'prop-types'
 import { useDispatch, useSelector } from 'react-redux'
 import MuiAlert from '@material-ui/lab/Alert'
-import PropTypes from 'prop-types'
 import { ACCENT_COLOR, TEXT_DIMMED_COLOR } from '../../core/values/colors'
 import {
+  userClosedErrorAlert,
   addTeamStarted,
   isFailedSelector,
   isFetchingSelector,
   isFulfilledSelector,
   setIsFulfilledFalse,
-  userClosedErrorAlert,
+  setRefresh,
 } from '../HomePage/homeSlice'
 
 export const useStyles = makeStyles((theme) => ({
@@ -75,6 +76,8 @@ export const useStyles = makeStyles((theme) => ({
 }))
 
 const AddTeamModal = (props) => {
+  const classes = useStyles()
+
   const { isModalOpen, setIsModalOpen } = props
 
   const isFulfilled = useSelector(isFulfilledSelector)
@@ -110,6 +113,7 @@ const AddTeamModal = (props) => {
   if (isFulfilled) {
     setIsModalOpen(false)
     dispatch(setIsFulfilledFalse())
+    dispatch(setRefresh())
   }
 
   const handleCloseAlert = (e, reason) => {
@@ -119,7 +123,6 @@ const AddTeamModal = (props) => {
     dispatch(userClosedErrorAlert())
   }
 
-  const classes = useStyles()
   return (
     <>
       <Modal
