+++ src/static/css/index.scss
@@ -16,16 +16,15 @@ $SUCCESS_COLOR: #A8ED8E;
 body {
   margin: 0;
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-  'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-  sans-serif;
+    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
+    sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
-  ::-webkit-scrollbar {display:none;}
 }
 
 code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
-  monospace;
+    monospace;
 }
 
 .MuiPaper-elevation1 {
@@ -40,7 +39,6 @@ a {
 .paperBtn {
   width: 100%;
   padding: 0 !important;
-
   .MuiPaper-root {
     width: 100%;
     text-align: left;
@@ -55,16 +53,13 @@ a {
 
 .addMemberBtn {
   width: 100%;
-
   .MuiTypography-root {
     text-transform: none;
   }
-
   &:hover {
     background: $ACCENT3_DIMMED_COLOR !important;
   }
 }
-
 .rounded {
   border-radius: 20px !important;
 }
